convert notifsapp to isomorphic single-page app
For https://github.com/shurcooL/home/issues/40.
