--- internal/exp/app/notifications/assets/doc.go
@@ -1,4 +1,0 @@
-//go:generate go run -tags=dev assets_gen.go
-
-// Package assets contains assets for notifications app.
-package assets
