+++ internal/exp/app/notifsapp/backend.go
@@ -1,0 +1,8 @@
+// +build !js !wasm
+// +build go1.14
+
+package notifsapp
+
+import "context"
+
+func (*app) SetupPage(context.Context, interface{}) {}
